+++ src/@locale/demo/metrics/index.js
@@ -1,7 +1,9 @@
+import barList     from './barList' ;
 import categoryBar from './categoryBar' ;
 
 const metrics =
 {
+    barList ,
     categoryBar ,
 } ;